Complete class names
in this commit only for "use" statement.

- in the main script we select class loader
which is an implementation of ClassLoaderInterface.
TODO: load NullClassLoader if using ComposerClassLoader does not make sense

- using getAbsoluteClassesPaths($pattern) we can query PHPID instance
about list of class that matches given pattern

- class loader can be reloaded. Now the only trigger is sent from the
PHPID::index() body right after executing composer dumpautoload.

Problem is that index() is called on the child process
but we need to update the loader of the parent PHPID process.
I resolved it using socket pair
to communication between these processes.

Method matchPattern() was moved up to RpcServer
because it is now needed in both subclasses.