StringBasedPHPFileInfo takes over nsuse() implementation
It is the place to create method that can help what to do
with a newly inserted alias i.e. if we have to add new import
or rename that alias because of conflict with another one.

The internals specific for processing each interesting line type
was divided into separate private methods,
so the logic of scanning the whole file
is more clear.

$specific_alias was removed