phpcd#completeDone()
- currently the method works for classes only
(i.e. if kind=='c' in completed item). It checks
if there is a need to fix the inserted alias and
to put new path to imports list

- added autocommand that calls this method.

TODO:
- if the server returns more than one suggestion,
then we need prompt user for select one of them.
Currently the first is selected without any interaction.

- if the alias fixed by server is not null (i.e. if is modified)
we need prompt user for edit it.
Currently the suggestion from server is returned.